export duplicated data select inputs into new component
